🔧 run_service.py 분리: heartbeats / discord_runner / work_order_executor_runner
## 변경 이유
- `runtime/run_service.py` 가 1387 LOC + 책임 5종 이상 (heartbeat
  / discord_runtime / runtime_orchestration / github_workflow /
  state_persistence) 으로 `code_audit` 의 split_now 위반.
- 한 파일에 SIGTERM dispatch / discord gateway / member bot / work
  order executor / supervisor builder / autonomy producer / coding
  executor bundle 이 다 섞여 있어 한 곳을 고치면 다른 곳 회귀 위험이
  높았다.

## 주요 변경 사항
- `runtime/heartbeats.py` 신설 (94 LOC)
  * `heartbeat_loop` — 기존 `_heartbeat_loop` 본체. shutdown_event
    aware, record 실패 시 parent loop 죽이지 않음.
  * `record_graceful_disable` — token 누락/placeholder 시 status
    surface 가 "operator disabled" 로 보이게 하는 best-effort
    heartbeat stamp.
- `runtime/discord_runner.py` 신설 (300 LOC)
  * `run_discord_gateway` — engineering gateway 단일 진입 + SIGTERM
    aware 종료 + heartbeat.
  * `run_discord_member_bot` — role-별 member bot 단일 진입 + token
    누락/placeholder graceful disable + heartbeat.
  * `install_role_runner_dispatch_for_run_service` — gateway boot
    이전 role-runner dispatcher 를 publish 하는 shim. legacy bot.py
    의 installer 와 동일한 trace stderr 출력.
- `runtime/work_order_executor_runner.py` 신설 (155 LOC)
  * `run_github_work_order_executor` — github_work_order 큐 단일
    consumer. live GitHub App env 있으면 GithubWriter, 없으면 None
    (SKIPPED_NO_WRITER audit). startup recovery hook 으로
    requeue_no_repo_failures 자동 실행.
  * `maybe_build_live_github_client` — 3 env 키 (APP_ID /
    INSTALLATION_ID / PRIVATE_KEY_PATH) 모두 set 시에만 live client
    build.
- `runtime/run_service.py` (1387 → 980 LOC)
  * 추출된 함수들을 `_*_impl` alias 로 import + module-level alias
    binding 유지. 기존 tests / external callers 호환.

## 비고
- 본 split 으로 `governance/code_audit` 가 run_service.py 를
  split_now 위반에서 제거 (980 LOC → warn). SPLIT_NOW_PENDING 에서도
  제거 완료.
- 다른 split_now 파일 (status.py, collector.py 등) 은 SPLIT_NOW_PENDING
  에 deadline 명시 — 후속 PR 에서 동일 패턴으로 분해.
- 전체 5079 test 통과 — runtime / governance / discord / engineering
  모두 회귀 없음.
